third_party/WebKit/LayoutTests/shadow-dom/event-post-process.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/shadow-dom.js"></script>

<div id="test1">
  <div id="d1">
    <div id="target"></div>
  </div>
</div>

<script>
test(() => {
  let n = createTestTree(test1);
  let event = new Event('my-event', { bubbles: true, composed: true });
  let log = dispatchEventWithLog(n, n.target, event);
  assert_equals(event.target, n.target);
  assert_equals(event.eventPhase, 0);
  assert_equals(event.currentTarget, null);
  assert_equals(event.composedPath().length, 0);
}, 'Event Path without ShadowRoots (composed: true).');
</script>

<script>
test(() => {
  let n = createTestTree(test1);
  let event = new Event('my-event', { bubbles: true, composed: false });
  let log = dispatchEventWithLog(n, n.target, event);
  assert_equals(event.target, n.target);
  assert_equals(event.eventPhase, 0);
  assert_equals(event.currentTarget, null);
  assert_equals(event.composedPath().length, 0);
}, 'Event Path without ShadowRoots (composed: false).');
</script>

<div id="test2">
  <div id="host">
    <template id="sr" data-mode="open">
      <div id="target"></div>
    </template>
  </div>
</div>

<script>
test(() => {
  let n = createTestTree(test2);
  let event = new Event('my-event', { bubbles: true, composed: true });
  let log = dispatchEventWithLog(n, n.target, event);
  assert_equals(event.target, n.target);
  assert_equals(event.eventPhase, 0);
  assert_equals(event.currentTarget, null);
  assert_equals(event.composedPath().length, 0);
}, 'Event Path with an open ShadowRoot (composed: true).');
</script>

<script>
test(() => {
  let n = createTestTree(test2);
  let event = new Event('my-event', { bubbles: true, composed: true });
  let log = dispatchEventWithLog(n, n.target, event);
  assert_equals(event.target, n.target);
  assert_equals(event.eventPhase, 0);
  assert_equals(event.currentTarget, null);
  assert_equals(event.composedPath().length, 0);
}, 'Event Path with an open ShadowRoot (composed: false).');
</script>

<div id="test3">
  <div id="host">
    <template id="sr" data-mode="closed">
      <div id="target"></div>
    </template>
  </div>
</div>

<script>
test(() => {
  let n = createTestTree(test3);
  let event = new Event('my-event', { bubbles: true, composed: true });
  let log = dispatchEventWithLog(n, n.target, event);
  assert_equals(event.target, n.host);
  assert_equals(event.eventPhase, 0);
  assert_equals(event.currentTarget, null);
  assert_equals(event.composedPath().length, 0);
}, 'Event Path with a closed ShadowRoot (composed: true).');
</script>

<script>
test(() => {
  let n = createTestTree(test3);
  let event = new Event('my-event', { bubbles: true, composed: true });
  let log = dispatchEventWithLog(n, n.target, event);
  assert_equals(event.target, n.target);
  assert_equals(event.eventPhase, 0);
  assert_equals(event.currentTarget, null);
  assert_equals(event.composedPath().length, 0);
}, 'Event Path with a closed ShadowRoot (composed: false).');
</script>
